+++ Sources/Sentry/include/SentryTracer.h
@@ -4,9 +4,19 @@
 NS_ASSUME_NONNULL_BEGIN
 
 @class SentryHub, SentryTransactionContext, SentryTraceHeader, SentryTraceContext,
-    SentryDispatchQueueWrapper;
+    SentryDispatchQueueWrapper, SentryTracer;
 
 static NSTimeInterval const SentryTracerDefaultTimeout = 3.0;
+
+@protocol SentryTracerDelegate
+
+/**
+ * Return the active span of given tracer.
+ * This function is used to determine which span will be used to create a new child.
+ */
+- (nullable id<SentrySpan>)activeSpanForTracer:(SentryTracer *)tracer;
+
+@end
 
 @interface SentryTracer : NSObject <SentrySpan>
 
@@ -35,6 +45,11 @@ static NSTimeInterval const SentryTracerDefaultTimeout = 3.0;
  */
 @property (readonly) BOOL isFinished;
 
+#if SENTRY_TARGET_PROFILING_SUPPORTED
+/** Whether the profiler is currently running. */
+@property (assign, readonly) BOOL isProfiling;
+#endif // SENTRY_TARGET_PROFILING_SUPPORTED
+
 @property (nullable, nonatomic, copy) void (^finishCallback)(SentryTracer *);
 
 /**
@@ -58,6 +73,11 @@ static NSTimeInterval const SentryTracerDefaultTimeout = 3.0;
  All the spans that where created with this tracer but rootSpan.
  */
 @property (nonatomic, readonly) NSArray<id<SentrySpan>> *children;
+
+/*
+ * A delegate that provides extra information for the transaction.
+ */
+@property (nullable, nonatomic, weak) id<SentryTracerDelegate> delegate;
 
 /**
  * Init a SentryTracer with given transaction context and hub and set other fields by default