+++ Tests/SentryTests/Performance/SentryTracerTests.swift
@@ -2,6 +2,15 @@ import XCTest
 
 class SentryTracerTests: XCTestCase {
     
+    private class TracerDelegate: SentryTracerDelegate {
+        
+        var activeSpan: Span?
+        
+        func activeSpan(for tracer: SentryTracer) -> Span? {
+            return activeSpan
+        }
+    }
+    
     private class Fixture {
         let client: TestClient
         let hub: TestHub
@@ -370,6 +379,53 @@ class SentryTracerTests: XCTestCase {
         
         let transaction = fixture.hub.capturedEventsWithScopes.first!.event as! Transaction
         assertAppStartsSpanAdded(transaction: transaction, startType: "Cold Start", operation: fixture.appStartColdOperation, appStartMeasurement: appStartMeasurement)
+    }
+    
+    func test_startChildWithDelegate() {
+        let delegate = TracerDelegate()
+        
+        let sut = fixture.getSut()
+        sut.delegate = delegate
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        
+        delegate.activeSpan = child
+        
+        let secondChild = sut.startChild(operation: fixture.transactionOperation)
+        
+        XCTAssertEqual(secondChild.context.parentSpanId, child.context.spanId)
+    }
+    
+    func test_startChildWithDelegate_ActiveNotChild() {
+        let delegate = TracerDelegate()
+        
+        let sut = fixture.getSut()
+        sut.delegate = delegate
+        
+        delegate.activeSpan = SentryTracer(transactionContext: TransactionContext(name: fixture.transactionName, operation: fixture.transactionOperation), hub: nil)
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        
+        let secondChild = sut.startChild(operation: fixture.transactionOperation)
+        
+        XCTAssertEqual(secondChild.context.parentSpanId, sut.context.spanId)
+        XCTAssertEqual(secondChild.context.parentSpanId, child.context.parentSpanId)
+    }
+    
+    func test_startChildWithDelegate_SelfIsActive() {
+        let delegate = TracerDelegate()
+        
+        let sut = fixture.getSut()
+        sut.delegate = delegate
+        
+        delegate.activeSpan = sut
+        
+        let child = sut.startChild(operation: fixture.transactionOperation)
+        
+        let secondChild = sut.startChild(operation: fixture.transactionOperation)
+        
+        XCTAssertEqual(secondChild.context.parentSpanId, sut.context.spanId)
+        XCTAssertEqual(secondChild.context.parentSpanId, child.context.parentSpanId)
     }
     
     func testAddWarmAppStartMeasurement_PutOnNextAutoUITransaction() {