+++ Tests/SentryTests/SentryCrash/TestSentryCrashWrapper.m
@@ -10,9 +10,11 @@
     instance.internalActiveDurationSinceLastCrash = NO;
     instance.internalActiveDurationSinceLastCrash = 0;
     instance.internalIsBeingTraced = NO;
+    instance.internalIsSimulatorBuild = NO;
     instance.internalIsApplicationInForeground = YES;
     instance.installAsyncHooksCalled = NO;
     instance.closeCalled = NO;
+    instance.internalFreeMemory = 0;
     return instance;
 }
 
@@ -29,6 +31,11 @@
 - (BOOL)isBeingTraced
 {
     return self.internalIsBeingTraced;
+}
+
+- (BOOL)isSimulatorBuild
+{
+    return self.internalIsSimulatorBuild;
 }
 
 - (BOOL)isApplicationInForeground
@@ -49,6 +56,11 @@
 - (NSDictionary *)systemInfo
 {
     return @{};
+}
+
+- (uint64_t)freeMemory
+{
+    return self.internalFreeMemory;
 }
 
 @end